tools/tutorials/fleet_management/vehicle-data-sender/requirements.txt
paho-mqtt == 2.1.0
